portail,login,registration

+++ app/src/main/java/com/example/centre_formation/MainActivity.java
@@ -7,7 +7,6 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.view.GravityCompat;
 import androidx.drawerlayout.widget.DrawerLayout;
 
-import android.app.Activity;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
@@ -15,26 +14,22 @@ import android.view.MenuItem;
 import android.widget.Button;
 import android.widget.TextView;
 
-import com.example.centre_formation.activity.Login;
 import com.example.centre_formation.activity.Registration;
 import com.example.centre_formation.database.AppDataBase;
 import com.example.centre_formation.entity.User;
+import com.example.centre_formation.fragment.LoginFragment;
 import com.google.android.material.navigation.NavigationView;
 import com.google.gson.Gson;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class MainActivity extends AppCompatActivity {
-Activity activity;
     private AppDataBase database;
     DrawerLayout drawerLayout;
     NavigationView navigationView;
     ActionBarDrawerToggle actionBarDrawerToggle;
     SharedPreferences shared;
-    TextView user;
+    public static final String PREF = "pref";
+    TextView userEmailInHeader, userNameInHeader;
     Button logout;
-    SharedPreferences myPref;
 
     @Override
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
@@ -57,11 +52,14 @@ Activity activity;
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
+        shared = getSharedPreferences(PREF, MODE_PRIVATE);
+        SharedPreferences.Editor editor = shared.edit();
+
         database = AppDataBase.getAppDatabase(this);
 
         drawerLayout = findViewById(R.id.drawer);
         navigationView = findViewById(R.id.navviewInMain);
-        actionBarDrawerToggle = new ActionBarDrawerToggle(this,this.drawerLayout,R.string.open,R.string.close);
+        actionBarDrawerToggle = new ActionBarDrawerToggle(this, this.drawerLayout, R.string.open, R.string.close);
         drawerLayout.addDrawerListener(actionBarDrawerToggle);
         actionBarDrawerToggle.syncState();
         ActionBar actionBar = getSupportActionBar();
@@ -71,44 +69,25 @@ Activity activity;
         navigationView.setNavigationItemSelectedListener(new NavigationView.OnNavigationItemSelectedListener() {
             @Override
             public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-                switch (item.getItemId()){
-
+                if (item.getItemId() == R.id.loginInMenu) {
+                    getSupportFragmentManager().beginTransaction().replace(R.id.fragment_container, new LoginFragment()).commit();
+                    drawerLayout.closeDrawer(GravityCompat.START);
+                    return true;
                 }
+                if (item.getItemId() == R.id.logoutInMenu) {
+                    editor.clear();
+                    editor.commit();
+                    Intent intent = getIntent();
+                    finish();
+                    startActivity(intent);
+                    drawerLayout.closeDrawer(GravityCompat.START);
+                    return true;
+                }
+                drawerLayout.closeDrawer(GravityCompat.START);
                 return false;
             }
         });
 
-        Button login = findViewById(R.id.goToLogin);
-        Button register = findViewById(R.id.goToRegister);
-        login.setOnClickListener(e -> {
-            Intent intent = new Intent(this, Login.class);
-            startActivity(intent);
-        });
-        register.setOnClickListener(e -> {
-            Intent intent2 = new Intent(this, Registration.class);
-            startActivity(intent2);
-        });
-
-
-        shared = getSharedPreferences(Login.PREF, MODE_PRIVATE);
-        user = findViewById(R.id.user);
-        String userJson = shared.getString("connectedUser", "null");
-        Gson gson = new Gson();
-        if (!userJson.equals("null")) {
-            User userJS = gson.fromJson(userJson, User.class);
-            user.setText(userJS.getFirstName());
-        }
-
-        myPref = getSharedPreferences(Login.PREF, MODE_PRIVATE);
-        SharedPreferences.Editor editor = myPref.edit();
-        logout = findViewById(R.id.logoutmain);
-        logout.setOnClickListener(e -> {
-            editor.clear();
-            editor.commit();
-            Intent intent = getIntent();
-            finish();
-            startActivity(intent);
-        });
 
     }
